requirements.txt
fastapi
uvicorn
websockets
pillow
google-genai